implememted the use of the masterservice to switch on and off other services
there are some workarounds due to incomplete firmware yet.
For example, we rely on a high reconnection delay because the user needs to "couple" calliope and ipad,
and because the current firmware does not send a response after writing the "services" characteristic,
but restarts immediately. So we need to wait for the timeout.